Switch to Signed Sensor Range, Update Arm Inversion
One previous issue was that sometimes the arm would think it was at 359 degrees rather than 0, so using the signed -180 to 180 range ensures that the value will be within a range of 4 degrees off rather than a full rotation. 

The arm inversion was also updated to match that of the sensor. Positive voltage is forward direction